api.support: move helper methods to own module
