+++ src/lando/api/legacy/api/__init__.py
@@ -1,3 +1,3 @@
-from lando.api.legacy.api import stacks, support
+from lando.api.legacy.api import stacks
 
 __all__ = ["stacks", "support"]
